+++ wizard/common.go
@@ -1,8 +1,17 @@
 package wizard
 
 import (
+	"bufio"
+	"os"
+	"regexp"
+	"strings"
+
 	"github.com/robocorp/rcc/common"
 	"github.com/robocorp/rcc/pretty"
+)
+
+var (
+	namePattern = regexp.MustCompile("^[\\w-]*$")
 )
 
 type WizardFn func([]string) error
@@ -18,4 +27,29 @@ func firstOf(arguments []string, missing string) string {
 		return arguments[0]
 	}
 	return missing
+}
+
+func note(message string) {
+	common.Stdout("%s! %s%s%s\n", pretty.Red, pretty.White, message, pretty.Reset)
+}
+
+func ask(question, defaults string, validator *regexp.Regexp, erratic string) (string, error) {
+	for {
+		common.Stdout("%s? %s%s %s[%s]:%s ", pretty.Green, pretty.White, question, pretty.Grey, defaults, pretty.Reset)
+		source := bufio.NewReader(os.Stdin)
+		reply, err := source.ReadString(newline)
+		common.Stdout("\n")
+		if err != nil {
+			return "", err
+		}
+		reply = strings.TrimSpace(reply)
+		if !validator.MatchString(reply) {
+			common.Stdout("%s%s%s\n\n", pretty.Red, erratic, pretty.Reset)
+			continue
+		}
+		if len(reply) == 0 {
+			return defaults, nil
+		}
+		return reply, nil
+	}
 }